Fix color converter, split into multiple files
