+++ listener_clients/embedded-listener/src/user_config_override_sample.h
@@ -1,1 +1,12 @@
 /* Place your custom build config here */
+
+// Uncomment this if you want to use the static IP
+/*
+#undef USE_STATIC_IP
+#define USE_STATIC_IP true
+
+#define STATIC_IP_ADDR IPAddress(192,168,0,99)
+#define GATEWAY_IP_ADDR IPAddress(192,168,0,1)
+#define SUBNET_ADDR IPAddress(255,255,255,0)
+#define DNS_ADDR IPAddress(192,168,0,1)
+*/
