+++ listener_clients/embedded-listener/src/utils.h
@@ -1,0 +1,31 @@
+extern SocketIOclient socketIO;
+
+void convertColorToRGB(String hexstring, int & r, int & g, int & b)
+{
+  hexstring.replace("#", "");
+  long number = (long) strtol( &hexstring[1], NULL, 16);
+
+  r = number >> 16;
+  g = number >> 8 & 0xFF;
+  b = number & 0xFF;
+}
+
+void sendSocketEvent(String event_name, DynamicJsonDocument params)
+{
+  // create JSON message for Socket.IO (event)
+  DynamicJsonDocument doc(1024);
+  JsonArray array = doc.to<JsonArray>();
+
+  array.add(event_name);
+  array.add(params);
+
+  // JSON to String (serializion)
+  String output;
+  serializeJson(doc, output);
+
+  // Send event
+  socketIO.sendEVENT(output);
+
+  // Print JSON for debugging
+  Serial.println(output);
+}
